add hackathon material compilation thoughts
I'm not sure if I put the paragraph under the right section. These steps are what we have proposed to implement to organize the contributions. It can also be mentioned that this format could be a useful prerequisite for contributions to future hackathons, but I wasn't sure if this would be within the scope of this chapter/paper.